Improve contact history loading
- Weird cubSortByKey problem is not solved

+++ src/kernel/DEMOwnerQueryKernels.cu
@@ -38,7 +38,7 @@ __global__ void computeKE(deme::DEMDataDT* granData, size_t nOwnerBodies, double
 
 __global__ void inspectOwnerProperty(deme::DEMDataDT* granData,
                                      deme::DEMSimParams* simParams,
-                                     float* quantity,
+                                     double* quantity,
                                      deme::notStupidBool_t* not_in_region,
                                      size_t nOwnerBodies) {
     deme::bodyID_t myOwner = blockIdx.x * blockDim.x + threadIdx.x;
